move http gateway property to abstract class and test
